analise_combustivel_rmtc/requirements.txt
pandas
numpy
python-dotenv
click
sqlalchemy
psycopg2
geopandas
shapely
geojson